# backend/api/gen/siem/v1/correlation_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: siem/v1/correlation.proto

package siemv1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationCorrelationGetCorrelatedRequest = "/siem.v1.Correlation/GetCorrelatedRequest"
const OperationCorrelationListCorrelatedRequests = "/siem.v1.Correlation/ListCorrelatedRequests"

type CorrelationHTTPServer interface {
	GetCorrelatedRequest(context.Context, *GetCorrelatedRequestRequest) (*CorrelatedRequest, error)
	ListCorrelatedRequests(context.Context, *ListCorrelatedRequestsRequest) (*ListCorrelatedRequestsResponse, error)
}

func RegisterCorrelationHTTPServer(s *http.Server, srv CorrelationHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/correlated/{correlation_id}", _Correlation_GetCorrelatedRequest0_HTTP_Handler(srv))
	r.GET("/api/v1/correlated", _Correlation_ListCorrelatedRequests0_HTTP_Handler(srv))
}

func _Correlation_GetCorrelatedRequest0_HTTP_Handler(srv CorrelationHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetCorrelatedRequestRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCorrelationGetCorrelatedRequest)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCorrelatedRequest(ctx, req.(*GetCorrelatedRequestRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CorrelatedRequest)
		return ctx.Result(200, reply)
	}
}

func _Correlation_ListCorrelatedRequests0_HTTP_Handler(srv CorrelationHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListCorrelatedRequestsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCorrelationListCorrelatedRequests)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListCorrelatedRequests(ctx, req.(*ListCorrelatedRequestsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListCorrelatedRequestsResponse)
		return ctx.Result(200, reply)
	}
}

type CorrelationHTTPClient interface {
	GetCorrelatedRequest(ctx context.Context, req *GetCorrelatedRequestRequest, opts ...http.CallOption) (rsp *CorrelatedRequest, err error)
	ListCorrelatedRequests(ctx context.Context, req *ListCorrelatedRequestsRequest, opts ...http.CallOption) (rsp *ListCorrelatedRequestsResponse, err error)
}

type CorrelationHTTPClientImpl struct {
	cc *http.Client
}

func NewCorrelationHTTPClient(client *http.Client) CorrelationHTTPClient {
	return &CorrelationHTTPClientImpl{client}
}

func (c *CorrelationHTTPClientImpl) GetCorrelatedRequest(ctx context.Context, in *GetCorrelatedRequestRequest, opts ...http.CallOption) (*CorrelatedRequest, error) {
	var out CorrelatedRequest
	pattern := "/api/v1/correlated/{correlation_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCorrelationGetCorrelatedRequest))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *CorrelationHTTPClientImpl) ListCorrelatedRequests(ctx context.Context, in *ListCorrelatedRequestsRequest, opts ...http.CallOption) (*ListCorrelatedRequestsResponse, error) {
	var out ListCorrelatedRequestsResponse
	pattern := "/api/v1/correlated"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCorrelationListCorrelatedRequests))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}