# backend/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/auth/login:
        post:
            tags:
                - Auth
            description: 'Step 1: verify the password. Returns an MFA challenge, never an access token.'
            operationId: Auth_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
    /api/v1/auth/logout:
        post:
            tags:
                - Auth
            operationId: Auth_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LogoutResponse'
...
    /api/v1/auth/mfa:
        post:
            tags:
                - Auth
            description: 'Step 2: verify the TOTP code and issue the token pair.'
            operationId: Auth_VerifyMFA
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/VerifyMFARequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TokenResponse'
    /api/v1/auth/refresh:
        post:
            tags:
                - Auth
            description: Exchange a refresh token for a new pair. The presented refresh token is revoked.
            operationId: Auth_Refresh
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RefreshRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TokenResponse'
    /api/v1/correlated:
        get:
            tags:
                - Correlation
            operationId: Correlation_ListCorrelatedRequests
            parameters:
                - name: timeRange.from
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: timeRange.to
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: confidence
                  in: query
                  description: Optional filter; unspecified returns every confidence level.
                  schema:
                    enum:
                        - CONFIDENCE_UNSPECIFIED
                        - CONFIDENCE_HIGH
                        - CONFIDENCE_MEDIUM
                        - CONFIDENCE_LOW
                    type: string
                    format: enum
                - name: onlyDisagreements
                  in: query
                  description: Restricts the result to records where vendors reached conflicting verdicts.
                  schema:
                    type: boolean
                - name: minVendorCount
                  in: query
                  description: Restricts the result to genuine cross-vendor joins when set above 1.
                  schema:
                    type: integer
                    format: uint32
                - name: page.cursor
                  in: query
                  description: Opaque cursor from a previous PageResponse. Empty requests the first page.
                  schema:
                    type: string
                - name: page.limit
                  in: query
                  description: Server caps this at QUERY_MAX_RESULT_ROWS regardless of what is requested.
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListCorrelatedRequestsResponse'
    /api/v1/correlated/{correlationId}:
        get:
            tags:
                - Correlation
            operationId: Correlation_GetCorrelatedRequest
            parameters:
                - name: correlationId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CorrelatedRequest'
...
components:
    schemas:
...
        ClientInfo:
            type: object
            properties:
                ip:
                    type: string
                ipShared:
                    type: boolean
                    description: |-
                        True when the address looks like NAT, a proxy chain, or a carrier range. This
                         downgrades join confidence rather than silently weakening the match.
                asn:
                    type: integer
                    format: uint32
                country:
                    type: string
                userAgent:
                    type: string
            description: ClientInfo describes the requesting client as normalized across vendors.
...
        CorrelatedRequest:
            type: object
            properties:
                correlationId:
                    type: string
                windowStart:
                    type: string
                    format: date-time
                firstEventTime:
                    type: string
                    format: date-time
                lastEventTime:
                    type: string
                    format: date-time
                client:
                    $ref: '#/components/schemas/ClientInfo'
                request:
                    $ref: '#/components/schemas/RequestInfo'
                vendorVerdicts:
                    type: array
                    items:
                        $ref: '#/components/schemas/VendorVerdict'
                    description: |-
                        One entry per contributing vendor, so a disagreement can be read directly off
                         the record rather than inferred from a flattened summary.
                vendorCount:
                    type: integer
                    description: '1 is normal, not an error: plenty of hostnames sit behind a single vendor (FR-016).'
                    format: uint32
                eventIds:
                    type: array
                    items:
                        type: string
                    description: Ids of the contributing events, for drilling through to the raw payloads (FR-024).
                combinedOutcome:
                    enum:
                        - VERDICT_UNSPECIFIED
                        - VERDICT_ALLOWED
                        - VERDICT_BLOCKED
                        - VERDICT_CHALLENGED
                        - VERDICT_RATE_LIMITED
                        - VERDICT_MONITORED
                        - VERDICT_UNKNOWN
                    type: string
                    description: The most restrictive verdict any vendor reached.
                    format: enum
                hasDisagreement:
                    type: boolean
                disagreementKind:
                    enum:
                        - DISAGREEMENT_KIND_UNSPECIFIED
                        - DISAGREEMENT_KIND_NONE
                        - DISAGREEMENT_KIND_ALLOW_VS_BLOCK
                        - DISAGREEMENT_KIND_ALLOW_VS_CHALLENGE
                        - DISAGREEMENT_KIND_SCORE_CONFLICT
                    type: string
                    format: enum
                joinSignals:
                    type: array
                    items:
                        enum:
                            - JOIN_SIGNAL_UNSPECIFIED
                            - JOIN_SIGNAL_VENDOR_REQUEST_ID
                            - JOIN_SIGNAL_IP_HOST_PATH_METHOD
                            - JOIN_SIGNAL_TIME_WINDOW
                        type: string
                        format: enum
                joinTier:
                    type: integer
                    description: 1 = exact shared request id, 2 = heuristic match.
                    format: uint32
                confidence:
                    enum:
                        - CONFIDENCE_UNSPECIFIED
                        - CONFIDENCE_HIGH
                        - CONFIDENCE_MEDIUM
                        - CONFIDENCE_LOW
                    type: string
                    format: enum
                candidateCount:
                    type: integer
                    description: Above 1, several events competed for the join and the chosen one may be wrong.
                    format: uint32
                version:
                    type: string
                    description: |-
                        Increments each time a late arrival amends the record. The correlation_id never
                         changes, so a bookmarked record stays findable.
                amended:
                    type: boolean
            description: CorrelatedRequest is one request as every vendor that saw it reported it.
...
        ListCorrelatedRequestsResponse:
            type: object
            properties:
                requests:
                    type: array
                    items:
                        $ref: '#/components/schemas/CorrelatedRequest'
                page:
                    $ref: '#/components/schemas/PageResponse'
...
        LoginRequest:
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
        LoginResponse:
            type: object
            properties:
                mfaChallengeToken:
                    type: string
                    description: Short-lived token proving the password step succeeded. Grants nothing else.
                mfaEnrolmentRequired:
                    type: boolean
                    description: True on first login, when the client must render the enrolment QR code.
                mfaProvisioningUri:
                    type: string
                    description: Populated only during enrolment, and never again afterwards.
        LogoutRequest:
            type: object
            properties:
                refreshToken:
                    type: string
        LogoutResponse:
            type: object
            properties: {}
...
        PageResponse:
            type: object
            properties:
                nextCursor:
                    type: string
                    description: Empty when there are no further pages.
                total:
                    type: string
                    description: |-
                        Approximate match count. Exact counts over hundreds of millions of rows cannot
                         meet the latency budget, so the API states when a count is an estimate rather
                         than implying precision it does not have.
                totalIsEstimate:
                    type: boolean
...
        RefreshRequest:
            type: object
            properties:
                refreshToken:
                    type: string
...
        RequestInfo:
            type: object
            properties:
                host:
                    type: string
                path:
                    type: string
                query:
                    type: string
                method:
                    type: string
                status:
                    type: integer
                    format: uint32
            description: RequestInfo describes the HTTP request as normalized across vendors.
...
        TokenResponse:
            type: object
            properties:
                accessToken:
                    type: string
                refreshToken:
                    type: string
                expiresAt:
                    type: string
                    format: date-time
                user:
                    $ref: '#/components/schemas/UserProfile'
...
        UserProfile:
            type: object
            properties:
                userId:
                    type: string
                email:
                    type: string
                role:
                    type: string
                    description: admin | analyst | auditor | ingest_only
                tenantId:
                    type: string
                tenantName:
                    type: string
                mfaEnabled:
                    type: boolean
                lastLoginAt:
                    type: string
                    format: date-time
        VendorVerdict:
            type: object
            properties:
                vendor:
                    enum:
                        - VENDOR_UNSPECIFIED
                        - VENDOR_CLOUDFLARE
                        - VENDOR_F5
                        - VENDOR_DATADOME
                    type: string
                    format: enum
                verdict:
                    enum:
                        - VERDICT_UNSPECIFIED
                        - VERDICT_ALLOWED
                        - VERDICT_BLOCKED
                        - VERDICT_CHALLENGED
                        - VERDICT_RATE_LIMITED
                        - VERDICT_MONITORED
                        - VERDICT_UNKNOWN
                    type: string
                    format: enum
                verdictReason:
                    type: string
                ruleId:
                    type: string
                ruleIds:
                    type: array
                    items:
                        type: string
                score:
                    type: number
                    description: Absent when the vendor does not score requests.
                    format: float
                scoreKind:
                    type: string
                eventId:
                    type: string
                eventTime:
                    type: string
                    format: date-time
            description: VendorVerdict is one vendor's contribution to a correlated request.
...
        VerifyMFARequest:
            type: object
            properties:
                mfaChallengeToken:
                    type: string
                code:
                    type: string
...